requirements.txt
streamlit==1.29.0
pandas==2.1.4
openpyxl==3.1.2
plotly==5.18.0
numpy==1.26.2
pillow==10.1.0
